requirements.txt
fastapi
uvicorn[standard]  
sqlmodel
psycopg2-binary     
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic-settings
python-json-logger
bcrypt==3.2.0
fastapi-pagination
twilio
pydantic[email]
alembic
pytz
Pillow>=9.0.0
boto3
openpyxl>=3.0.0
pandas